+++ src/context.rs
@@ -4,6 +4,9 @@ use std::collections::BTreeMap;
 use serde::ser::Serialize;
 use serde_json::value::{to_value, Map, Value};
 
+use crate::errors::{Result as TeraResult, Error};
+
+
 /// The struct that holds the context of a template rendering.
 ///
 /// Light wrapper around a `BTreeMap` for easier insertions of Serializable
@@ -15,7 +18,7 @@ pub struct Context {
 
 impl Context {
     /// Initializes an empty context
-    pub fn new() -> Context {
+    pub fn new() -> Self {
         Context { data: BTreeMap::new() }
     }
 
@@ -53,6 +56,28 @@ impl Context {
             m.insert(key, value);
         }
         Value::Object(m)
+    }
+
+    /// Takes a serde-json `Value` and convert it into a `Context` with no overhead/cloning
+    pub fn from_value(obj: Value) -> TeraResult<Self> {
+        match obj {
+            Value::Object(m) => {
+                let mut data = BTreeMap::new();
+                for (key, value) in m {
+                    data.insert(key, value);
+                }
+                Ok(Context { data })
+            },
+            _ => Err(Error::msg("Creating a Context from a Value/Serialize requires it being a JSON object"))
+        }
+    }
+
+    /// Takes something that impl Serialize and create a context with it.
+    /// Meant to be used if you have a hashmap or a struct and don't want to insert values
+    /// one by one in the context
+    pub fn from_serialize(value: impl Serialize) -> TeraResult<Self> {
+        let obj = to_value(value).map_err(Error::json)?;
+        Context::from_value(obj)
     }
 }
 
@@ -142,8 +167,11 @@ pub fn get_json_pointer(key: &str) -> String {
 mod tests {
     use super::*;
 
+    use std::collections::HashMap;
+    use serde_json::json;
+
     #[test]
-    fn test_extend() {
+    fn can_extend_context() {
         let mut target = Context::new();
         target.insert("a", &1);
         target.insert("b", &2);
@@ -154,5 +182,30 @@ mod tests {
         assert_eq!(*target.data.get("a").unwrap(), to_value(1).unwrap());
         assert_eq!(*target.data.get("b").unwrap(), to_value(3).unwrap());
         assert_eq!(*target.data.get("c").unwrap(), to_value(4).unwrap());
+    }
+
+    #[test]
+    fn can_create_context_from_value() {
+        let obj = json!({
+            "name": "bob",
+            "age": 25
+        });
+        let context_from_value = Context::from_value(obj).unwrap();
+        let mut context = Context::new();
+        context.insert("name", "bob");
+        context.insert("age", &25);
+        assert_eq!(context_from_value, context);
+    }
+
+    #[test]
+    fn can_create_context_from_impl_serialize() {
+        let mut map = HashMap::new();
+        map.insert("name", "bob");
+        map.insert("last_name", "something");
+        let context_from_serialize = Context::from_serialize(&map).unwrap();
+        let mut context = Context::new();
+        context.insert("name", "bob");
+        context.insert("last_name", "something");
+        assert_eq!(context_from_serialize, context);
     }
 }